Remove NonTaskCell
