+++ App/Tasks/TaskTypeEstimator.swift
@@ -16,7 +16,7 @@ class TaskTypeEstimator {
     func taskTypeAroundDate (_ date: Date, withSettings settings: Settings) -> TaskType {
 		
 		// Check if the date is around scrum time
-        var settingsScrumTime = gregorian.dateComponents(ymdhmsUnitFlags, from: settings.settingsTracking.scrumTime)
+        let settingsScrumTime = gregorian.dateComponents(ymdhmsUnitFlags, from: settings.settingsTracking.scrumTime)
 		
 		var comps = gregorian.dateComponents(ymdhmsUnitFlags, from: date)
 		comps.hour = settingsScrumTime.hour
@@ -30,7 +30,7 @@ class TaskTypeEstimator {
 		
 		
         // Check if the date is around lunch break
-        var settingsLunchTime = gregorian.dateComponents(ymdhmsUnitFlags, from: settings.settingsTracking.lunchTime)
+        let settingsLunchTime = gregorian.dateComponents(ymdhmsUnitFlags, from: settings.settingsTracking.lunchTime)
 		
 		comps.hour = settingsLunchTime.hour
 		comps.minute = settingsLunchTime.minute
