+++ Delivery/macOS/Screens/Tasks/cells/CellProtocol.swift
@@ -20,7 +20,7 @@ protocol CellProtocol {
     var color: NSColor {get set}
     var timeToolTip: String? {get set}
 	
-	var didEndEditingCell: ((_ cell: CellProtocol) -> ())? {get set}
+	var didClickEditCell: ((_ cell: CellProtocol) -> ())? {get set}
 	var didClickRemoveCell: ((_ cell: CellProtocol) -> ())? {get set}
 	var didClickAddCell: ((_ cell: CellProtocol) -> ())? {get set}
 	var didCopyContentCell: ((_ cell: CellProtocol) -> ())? {get set}
