--- Delivery/macOS/Screens/Tasks/cells/TaskCell/TaskCellTests.swift
@@ -1,39 +1,0 @@
-//
-//  TaskCellTests.swift
-//  Jirassic
-//
-//  Created by Baluta Cristian on 17/09/15.
-//  Copyright © 2015 Cristian Baluta. All rights reserved.
-//
-
-import XCTest
-@testable import Jirassic_no_cloud
-
-class TaskCellTests: XCTestCase {
-	
-//    func testSettingData() {
-//		
-//		let _cell = cell()
-//		RCLog(_cell)
-//		_cell.data = (dateStart: "dateStart", dateEnd: "dateEnd", issue: "IOS-01", notes: "notes")
-//		
-//		XCTAssert(_cell.data.dateStart == "dateStart", "")
-//		XCTAssert(_cell.data.dateEnd == "dateEnd", "")
-//		XCTAssert(_cell.data.issue == "IOS-01", "")
-//		XCTAssert(_cell.data.notes == "notes", "")
-//    }
-//	
-//	func cell() -> TaskCell {
-//		var views: NSArray?
-//		NSBundle.mainBundle().loadNibNamed("TaskCell", owner: self, topLevelObjects: &views)
-//		RCLog(views)
-//		RCLog(views?.objectAtIndex(0))
-//		RCLog(views?.objectAtIndex(1))
-//		if let c = views?.objectAtIndex(0) as? TaskCell {
-//			return c
-//		} else if let c = views?.objectAtIndex(1) as? TaskCell {
-//			return c
-//		}
-//		return TaskCell()
-//	}
-}
